added update user authorities by admin feature

+++ emles_oauth_server/src/main/java/com/emles/repository/AuthorityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.emles.model.Authority;
 
-public interface AuthorityRepository extends CrudRepository<Authority, Integer> {
+public interface AuthorityRepository extends CrudRepository<Authority, Long> {
     Authority findByAuthority(String authority);
 }
